+++ woocommerce-dashboard.php
@@ -4,6 +4,7 @@ add_filter( 'woocommerce_account_menu_items', 'thegift_woocommerce_sidebar_items
 function thegift_woocommerce_sidebar_items( $items ) {
   $reordered_items = array(
     'dashboard' => __('Enrolled Courses', 'woocommerce'),
+    'resources' => __('Additional Resources', 'woocommerce'),
     'edit-account' => __('Account Details', 'woocommerce'),
     'edit-address' => __('Addresses', 'woocommerce'),
     'orders' => __('Orders', 'woocommerce'),
@@ -20,4 +21,30 @@ function thegift_override_woo_dash( $located, $template_name, $args, $template_p
   }
   
   return $located;
+}
+
+add_action( 'init', 'thegift_add_woo_endpoints' );
+function thegift_add_woo_endpoints() {
+	add_rewrite_endpoint( 'resources', EP_PAGES );
+}
+
+add_action( 'woocommerce_account_resources_endpoint', 'thegift_woo_resources_endpoint_content' );
+function thegift_woo_resources_endpoint_content() {
+  $args = array(
+    'post_type' => 'resource',
+    'nopaging' => true,
+    'status' => 'publish'
+  );
+  $resources = get_posts( $args );
+  echo '<h2>Additional Resources</h2>';
+
+  if( $resources ){
+    include('templates/frontend/additional-resources.php');
+    wp_enqueue_style(
+      'woo-resources-stylesheet',
+      THEGIFT_PLUGIN_ROOT . '/stylesheets/additional-resources.css'
+    );
+  } else {  
+    echo '<p>There are no additional resources at this time. Check back again soon!</p>';
+  }
 }