+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## Version 1.2.0 - June 2, 2024
 - Create 'scale' type application question
+- Create Additional Resources in customer dashboard.
 
 ## Version 1.1.1 - May 31, 2024
 - Update Styles for Lessons